Move `IncrementalState` type from shelley to core
so it's accessible in EpochBoundary for era translation
