+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState/Types.hs
@@ -42,12 +42,12 @@ import Cardano.Ledger.Binary (
  )
 import Cardano.Ledger.Binary.Coders (Decode (From, RecD), decode, (<!))
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Credential (Credential (..), Ptr (..))
-import Cardano.Ledger.Crypto (Crypto)
+import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DPState (DPState)
 import Cardano.Ledger.EpochBoundary (
   SnapShots (..),
  )
+import Cardano.Ledger.IncrementalStake (IncrementalStake (..))
 import Cardano.Ledger.Keys (
   KeyHash (..),
   KeyPair, -- deprecated
@@ -64,9 +64,7 @@ import Control.DeepSeq (NFData)
 import Control.Monad.State.Strict (evalStateT)
 import Control.Monad.Trans (MonadTrans (lift))
 import Data.Default.Class (Default, def)
-import Data.Group (Group, invert)
 import Data.Map.Strict (Map)
-import qualified Data.Map.Strict as Map
 import GHC.Generics (Generic)
 import Lens.Micro (_1, _2)
 import NoThunks.Class (NoThunks (..))
@@ -173,46 +171,6 @@ instance (EraTxOut era, EraGovernance era) => ToCBOR (EpochState era) where
 
 instance (EraTxOut era, EraGovernance era) => FromCBOR (EpochState era) where
   fromCBOR = fromEraCBOR @era
-
--- =============================
-
--- | Incremental Stake, Stake along with possible missed coins from danging Ptrs.
---   Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr
---   does not have to point to anything until the epoch boundary, when we compute
---   rewards and aggregate staking information for ranking. This is unusual but legal.
---   In a non incremental system, we use whatever 'legal' Ptrs exist at the epoch
---   boundary. Here we are computing things incrementally, so we need to remember Ptrs
---   that might point to something by the time the epoch boundary is reached. When
---   the epoch boundary is reached we 'resolve' these pointers, to see if any have
---   become non-dangling since the time they were first used in the incremental computation.
-data IncrementalStake c = IStake
-  { credMap :: !(Map (Credential 'Staking c) Coin)
-  , ptrMap :: !(Map Ptr Coin)
-  }
-  deriving (Generic, Show, Eq, Ord, NoThunks, NFData)
-
-instance Crypto c => EncCBOR (IncrementalStake c) where
-  encCBOR (IStake st dangle) =
-    encodeListLen 2 <> encCBOR st <> encCBOR dangle
-
-instance Crypto c => DecShareCBOR (IncrementalStake c) where
-  type Share (IncrementalStake c) = Interns (Credential 'Staking c)
-  decShareCBOR credInterns =
-    decodeRecordNamed "Stake" (const 2) $ do
-      stake <- decShareCBOR (credInterns, mempty)
-      IStake stake <$> decCBOR
-
-instance Semigroup (IncrementalStake c) where
-  (IStake a b) <> (IStake c d) = IStake (Map.unionWith (<>) a c) (Map.unionWith (<>) b d)
-
-instance Monoid (IncrementalStake c) where
-  mempty = IStake Map.empty Map.empty
-
-instance Data.Group.Group (IncrementalStake c) where
-  invert (IStake m1 m2) = IStake (Map.map invert m1) (Map.map invert m2)
-
-instance Default (IncrementalStake c) where
-  def = IStake Map.empty Map.empty
 
 -- =============================
 
@@ -516,5 +474,3 @@ instance
   , ToExpr (GovernanceState era)
   ) =>
   ToExpr (UTxOState era)
-
-instance ToExpr (IncrementalStake c)
