+++ libs/cardano-ledger-core/src/Cardano/Ledger/IncrementalStake.hs
@@ -1,0 +1,79 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-redundant-constraints #-}
+
+module Cardano.Ledger.IncrementalStake (
+  IncrementalStake (..),
+) where
+
+import Cardano.Ledger.Binary (
+  DecCBOR (decCBOR),
+  DecShareCBOR (Share, decShareCBOR),
+  EncCBOR (encCBOR),
+  Interns,
+  decodeRecordNamed,
+  encodeListLen,
+ )
+import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Ledger.Credential (Credential (..), Ptr (..))
+import Cardano.Ledger.Crypto (Crypto)
+import Cardano.Ledger.Keys (
+  -- deprecated
+  KeyRole (..),
+ )
+import Cardano.Ledger.TreeDiff (ToExpr)
+import Control.DeepSeq (NFData)
+import Data.Default.Class (Default, def)
+import Data.Group (Group, invert)
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
+import GHC.Generics (Generic)
+import NoThunks.Class (NoThunks (..))
+
+-- | Incremental Stake, Stake along with possible missed coins from danging Ptrs.
+--   Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr
+--   does not have to point to anything until the epoch boundary, when we compute
+--   rewards and aggregate staking information for ranking. This is unusual but legal.
+--   In a non incremental system, we use whatever 'legal' Ptrs exist at the epoch
+--   boundary. Here we are computing things incrementally, so we need to remember Ptrs
+--   that might point to something by the time the epoch boundary is reached. When
+--   the epoch boundary is reached we 'resolve' these pointers, to see if any have
+--   become non-dangling since the time they were first used in the incremental computation.
+data IncrementalStake c = IStake
+  { credMap :: !(Map (Credential 'Staking c) Coin)
+  , ptrMap :: !(Map Ptr Coin)
+  }
+  deriving (Generic, Show, Eq, Ord, NoThunks, NFData)
+
+instance Crypto c => EncCBOR (IncrementalStake c) where
+  encCBOR (IStake st dangle) =
+    encodeListLen 2 <> encCBOR st <> encCBOR dangle
+
+instance Crypto c => DecShareCBOR (IncrementalStake c) where
+  type Share (IncrementalStake c) = Interns (Credential 'Staking c)
+  decShareCBOR credInterns =
+    decodeRecordNamed "Stake" (const 2) $ do
+      stake <- decShareCBOR (credInterns, mempty)
+      IStake stake <$> decCBOR
+
+instance Semigroup (IncrementalStake c) where
+  (IStake a b) <> (IStake c d) = IStake (Map.unionWith (<>) a c) (Map.unionWith (<>) b d)
+
+instance Monoid (IncrementalStake c) where
+  mempty = IStake Map.empty Map.empty
+
+instance Data.Group.Group (IncrementalStake c) where
+  invert (IStake m1 m2) = IStake (Map.map invert m1) (Map.map invert m2)
+
+instance Default (IncrementalStake c) where
+  def = IStake Map.empty Map.empty
+
+instance ToExpr (IncrementalStake c)
